refactor(vercelaisdk): drop the model-level handler
The model-level Handler shipped its own agentic tool loop (StreamModelWithTools).
With AgentHandler bridging at the agent layer — where the loop, tools, prompt,
and interceptors actually live — that second loop is dead weight and a source of
divergence. Remove it.

Gone: Handler/ServeHTTP, StreamModel, StreamModelWithTools + its turn helpers,
executeTools, ToolExecutor, the static tool writers, handleContentPart,
writeStreamEnd, mapFinishReason, and the WithSystem/WithTools/WithMaxTurns
options. convertMessages loses its always-empty system parameter.

handler.go is now the shared protocol primitives (message conversion, the SSE
EventWriter, the streamWriter span bookkeeping) that AgentHandler builds on. The
model-level handler_test.go and the conformance suite (which existed solely to
exercise that handler) are removed; the reusable convertMessages/reconstruction
unit tests move to convert_test.go. Agent-level conformance against the real ai
package is a tracked follow-up.
