--- adapter/vercelaisdk/uimessagestream/conformance/package.json
@@ -1,11 +1,0 @@
-{
-  "name": "aisdk-conformance",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "test": "node --test conformance.test.mjs"
-  },
-  "dependencies": {
-    "ai": "^7.0.6"
-  }
-}
